+++ backend/routes/setup.go
@@ -3,9 +3,10 @@ package routes
 import (
 	"github.com/gin-gonic/gin"
 	"github.com/kubestellar/ui/plugin/plugins"
+	"github.com/kubestellar/ui/services" 
 )
 
-func SetupRoutes(router *gin.Engine) {
+func SetupRoutes(router *gin.Engine, pluginService *services.PluginService) {
 	// Initialize all route groups
 	setupClusterRoutes(router)
 	setupDeploymentRoutes(router)
@@ -19,7 +20,14 @@ func SetupRoutes(router *gin.Engine) {
 	setupHelmRoutes(router)
 	setupGitHubRoutes(router)
 	setupDeploymentHistoryRoutes(router)
+
+	// Setup existing plugin routes
 	plugins.Pm.SetupPluginsRoutes(router)
+
+	// Setup new dynamic plugin management routes
+	if pluginService != nil {
+		SetupPluginRoutes(router)
+	}
 
 	setupAuthRoutes(router)
 	setupArtifactHubRoutes(router)
